[feature] Update SimilarityChecker::CalcLengthScore function - Fix the type casting error

+++ SimilarityChecker/SimilarityChecker.cpp
@@ -11,8 +11,8 @@ public:
 	int CalcLengthScore(const string& str1, const string& str2) {
 		int long_length = std::max(str1.size(), str2.size());
 		int short_length = std::min(str1.size(), str2.size());
-		int gap = long_length - short_length;
-		int result = (1.0 - double(gap / short_length)) * 60;
+		double gap = static_cast<double>(long_length - short_length);
+		int result = double(1.0 - (gap / double(short_length))) * 60.0;
 		return result;
 	}
 };
